(Lab3)Implementasi Template HTML sederhana berdasarkan fitur CRUD sekaligus final code

+++ iet_24782072_2026/main_app/views.py
@@ -1,6 +1,11 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Report
 from .forms import ReportForm
+
+
+# HOME
+def home(request):
+    return render(request, 'main_app/home.html')
 
 
 # CREATE
@@ -24,7 +29,7 @@ def report_list(request):
 
 # UPDATE
 def update_report(request, id):
-    report = Report.objects.get(id=id)
+    report = get_object_or_404(Report, id=id)
 
     if request.method == "POST":
         form = ReportForm(request.POST, instance=report)
@@ -40,5 +45,9 @@ def update_report(request, id):
 # DELETE
 def delete_report(request, id):
     report = Report.objects.get(id=id)
-    report.delete()
-    return redirect('report_list')
+
+    if request.method == 'POST':
+        report.delete()
+        return redirect('report_list')
+
+    return render(request, 'main_app/delete_report.html', {'report': report})
